@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,7 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=2.19.6 <3.0.0"
+  flutter: ^3.7.9
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -33,10 +34,10 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  go_router: ^6.0.0
+  go_router: ^6.5.5
   flutter_riverpod: ^2.1.1
-  dio: ^4.0.6
-  retrofit: ^3.3.1
+  dio: ^5.1.1
+  retrofit: ^4.0.1
   equatable: ^2.0.5
 
   shimmer: ^2.0.0
@@ -45,19 +46,18 @@ dependencies:
   advance_pdf_viewer:
     git:
       url: https://github.com/enveraltin/pdf_viewer
-  responsive_builder: ^0.4.3
+  responsive_builder: ^0.6.4
   expandable: ^5.0.1
-  lottie: ^2.1.0
-  path_provider: ^2.0.11
+  lottie: ^2.3.2
+  path_provider: ^2.0.14
   hive_flutter: ^1.1.0
   flutter_sticky_header: ^0.6.5
   freezed_annotation: ^2.2.0
   riverpod_annotation: ^1.0.6
-  flutter_downloader: ^1.9.1
+  flutter_downloader: ^1.10.2
   cached_network_image: ^3.2.3
-  flutter_expandable_fab: ^1.6.1
-
-
+  flutter_expandable_fab: ^1.7.1
+  json_annotation: ^4.8.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -71,8 +71,8 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^2.0.0
-  retrofit_generator: ^4.2.0
-  json_serializable: ^6.5.4
+  retrofit_generator: ^6.0.0+1
+  json_serializable: ^6.6.1
   change_app_package_name: ^1.1.0
 
 # For information on the generic Dart part of this file, see the
@@ -80,7 +80,6 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
